[AMI] authenticated registration endpoints (#101)

+++ app/controllers/registration.py
@@ -1,8 +1,10 @@
 import uuid
+from collections.abc import Sequence
 from typing import Annotated
 
 from advanced_alchemy.extensions.litestar import providers
 from litestar import Controller, Response, delete, get, post
+from litestar.di import Provide
 from litestar.exceptions import NotFoundException
 from litestar.params import Body
 from litestar.status_codes import (
@@ -14,23 +16,20 @@ from webpush import WebPushSubscription
 
 from app import models, schemas
 from app.services.registration import RegistrationService
-from app.services.user import UserService
+from app.services.user import provide_user
 
 
 class RegistrationController(Controller):
     dependencies = {
+        "current_user": Provide(provide_user),
         "registrations_service": providers.create_service_provider(RegistrationService),
-        "users_service": providers.create_service_provider(UserService),
-        "users_with_registrations_service": providers.create_service_provider(
-            UserService, load=[models.User.registrations]
-        ),
     }
 
     @post("/api/v1/users/registrations")
     async def register(
         self,
         registrations_service: RegistrationService,
-        users_service: UserService,
+        current_user: models.User,
         data: Annotated[
             schemas.RegistrationCreate,
             Body(
@@ -40,13 +39,12 @@ class RegistrationController(Controller):
         ],
     ) -> Response[schemas.Registration]:
         WebPushSubscription.model_validate(data.subscription)
-        user: models.User | None = await users_service.get_one_or_none(id=data.user_id)
-        if user is None:
-            raise NotFoundException(detail="User not found")
 
         existing_registration: (
             models.Registration | None
-        ) = await registrations_service.get_one_or_none(subscription=data.subscription, user=user)
+        ) = await registrations_service.get_one_or_none(
+            subscription=data.subscription, user=current_user
+        )
         if existing_registration is not None:
             # This registration already exists, don't duplicate it.
             return Response(
@@ -57,34 +55,40 @@ class RegistrationController(Controller):
             )
 
         registration: models.Registration = await registrations_service.create(
-            models.Registration(**data.model_dump())
+            models.Registration(user=current_user, **data.model_dump())
         )
         return Response(
             registrations_service.to_schema(registration, schema_type=schemas.Registration),
             status_code=HTTP_201_CREATED,
         )
 
-    @get("/api/v1/users/{user_id:uuid}/registrations")
+    @get("/api/v1/users/registrations")
     async def list_registrations(
         self,
         registrations_service: RegistrationService,
-        users_with_registrations_service: UserService,
-        user_id: uuid.UUID,
+        current_user: models.User,
     ) -> list[schemas.Registration]:
-        user: models.User | None = await users_with_registrations_service.get_one_or_none(
-            id=user_id
-        )
-        if user is None:
-            raise NotFoundException(detail="User not found")
         # We could do:
         # return registrations_service.to_schema(user.registrations, schema_type=schemas.Registration)
         # But it adds pagination.
         # For the moment, just return a list of dict
+        registrations: Sequence[models.Registration] = await registrations_service.list(
+            user=current_user,
+        )
         type_adapter = TypeAdapter(list[schemas.Registration])
-        return type_adapter.validate_python(user.registrations)
+        return type_adapter.validate_python(registrations)
 
     @delete("/api/v1/users/registrations/{registration_id:uuid}")
     async def unregister(
-        self, registrations_service: RegistrationService, registration_id: uuid.UUID
+        self,
+        registrations_service: RegistrationService,
+        current_user: models.User,
+        registration_id: uuid.UUID,
     ) -> None:
+        registration: models.Registration | None = await registrations_service.get_one_or_none(
+            id=registration_id,
+            user=current_user,
+        )
+        if registration is None:
+            raise NotFoundException(detail="Registration not found")
         await registrations_service.delete(registration_id)
